refactor(app): move ProtocolSetup and ProtocolDashboard
