+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -28,10 +28,10 @@ import { ConnectViaWifi } from '../pages/ConnectViaWifi'
 import { EmergencyStop } from '../pages/EmergencyStop'
 import { NameRobot } from '../pages/NameRobot'
 import { NetworkSetupMenu } from '../pages/NetworkSetupMenu'
-import { ProtocolSetup } from '../pages/ProtocolSetup'
+import { ProtocolSetup } from '../pages/ODD/ProtocolSetup'
 import { RobotDashboard } from '../pages/RobotDashboard'
 import { RobotSettingsDashboard } from '../pages/RobotSettingsDashboard'
-import { ProtocolDashboard } from '../pages/ProtocolDashboard'
+import { ProtocolDashboard } from '../pages/ODD/ProtocolDashboard'
 import { ProtocolDetails } from '../pages/ODD/ProtocolDetails'
 import { QuickTransferFlow } from '../organisms/QuickTransferFlow'
 import { QuickTransferDashboard } from '../pages/ODD/QuickTransferDashboard'
