+++ app/src/organisms/ProtocolSetupDeckConfiguration/index.tsx
@@ -32,7 +32,7 @@ import type {
   ModuleModel,
 } from '@opentrons/shared-data'
 import type { ModuleOnDeck } from '@opentrons/components'
-import type { SetupScreens } from '../../pages/ProtocolSetup'
+import type { SetupScreens } from '../../pages/ODD/ProtocolSetup'
 
 interface ProtocolSetupDeckConfigurationProps {
   cutoutId: CutoutId | null
