+++ app/src/pages/ODD/ProtocolDashboard/LongPressModal.tsx
@@ -12,13 +12,13 @@ import {
 } from '@opentrons/components'
 import { useCreateRunMutation } from '@opentrons/react-api-client'
 
-import { MAXIMUM_PINNED_PROTOCOLS } from '../../App/constants'
-import { SmallModalChildren } from '../../molecules/OddModal'
-import { useToaster } from '../../organisms/ToasterOven'
-import { getPinnedProtocolIds, updateConfigValue } from '../../redux/config'
+import { MAXIMUM_PINNED_PROTOCOLS } from '../../../App/constants'
+import { SmallModalChildren } from '../../../molecules/OddModal'
+import { useToaster } from '../../../organisms/ToasterOven'
+import { getPinnedProtocolIds, updateConfigValue } from '../../../redux/config'
 
 import type { UseLongPressResult } from '@opentrons/components'
-import type { Dispatch } from '../../redux/types'
+import type { Dispatch } from '../../../redux/types'
 
 interface LongPressModalProps {
   longpress: UseLongPressResult
