+++ app/src/pages/ODD/ProtocolDashboard/ProtocolCard.tsx
@@ -29,14 +29,14 @@ import {
 } from '@opentrons/react-api-client'
 import { deleteProtocol, deleteRun, getProtocol } from '@opentrons/api-client'
 
-import { SmallButton } from '../../atoms/buttons'
-import { OddModal } from '../../molecules/OddModal'
+import { SmallButton } from '../../../atoms/buttons'
+import { OddModal } from '../../../molecules/OddModal'
 import { LongPressModal } from './LongPressModal'
-import { formatTimeWithUtcLabel } from '../../resources/runs'
+import { formatTimeWithUtcLabel } from '../../../resources/runs'
 
 import type { UseLongPressResult } from '@opentrons/components'
 import type { ProtocolResource } from '@opentrons/shared-data'
-import type { OddModalHeaderBaseProps } from '../../molecules/OddModal/types'
+import type { OddModalHeaderBaseProps } from '../../../molecules/OddModal/types'
 
 const REFETCH_INTERVAL = 5000
