+++ app/src/pages/ODD/ProtocolDashboard/__tests__/ProtocolCard.test.tsx
@@ -9,9 +9,9 @@ import {
   useProtocolAnalysisAsDocumentQuery,
 } from '@opentrons/react-api-client'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
-import { useFeatureFlag } from '../../../redux/config'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
+import { useFeatureFlag } from '../../../../redux/config'
 import { ProtocolCard } from '../ProtocolCard'
 
 import type { NavigateFunction } from 'react-router-dom'
@@ -32,7 +32,7 @@ vi.mock('react-router-dom', async importOriginal => {
   }
 })
 vi.mock('@opentrons/react-api-client')
-vi.mock('../../../redux/config')
+vi.mock('../../../../redux/config')
 vi.mock('@opentrons/components', async importOriginal => {
   const actual = await importOriginal<typeof Chip>()
   return {
