+++ app/src/pages/ODD/ProtocolDashboard/utils.ts
@@ -1,6 +1,6 @@
 import type { ProtocolResource } from '@opentrons/shared-data'
 import type { RunData } from '@opentrons/api-client'
-import type { ProtocolsOnDeviceSortKey } from '../../redux/config/types'
+import type { ProtocolsOnDeviceSortKey } from '../../../redux/config/types'
 
 const DUMMY_FOR_NO_DATE_CASE = -8640000000000000
