+++ app/src/pages/ODD/ProtocolSetup/__tests__/ConfirmAttachedModal.test.tsx
@@ -2,10 +2,10 @@ import * as React from 'react'
 import { vi, it, describe, expect, beforeEach } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 
-import { i18n } from '../../../i18n'
-import { ConfirmAttachedModal } from '../../../pages/ProtocolSetup/ConfirmAttachedModal'
+import { i18n } from '../../../../i18n'
+import { ConfirmAttachedModal } from '../ConfirmAttachedModal'
 
 const mockOnCloseClick = vi.fn()
 const mockOnConfirmClick = vi.fn()
